+++ CHANGES.md
@@ -1,4 +1,5 @@
 - New Feature: Add a new page to view single donor details.
 - New Feature: Send e-mail notifications after users or donors registration.
+- Enhancement: Improve the user experience in managing pending users and donors.
 - Enhancement: Keep forms fields values after failed entities addition submissions.
 - Bug Fix: Add a check to avoid registering a user with an existing e-mail address.